added option to change maze parameters from outside the code

+++ ppi/MazeManager.h
@@ -12,7 +12,7 @@ public:
 	void generateDFS();
 	void removeOneRandWall(int x,int y);
 	void draw(sf::RenderWindow& window);
-	void braidmaze(float chance);
+	void braidmaze(int chance);
 	void markPoints() {
 		maze[player.first][player.second].setFillColor(sf::Color::Green);
 		maze[height - 1][width - 1].setFillColor(sf::Color::Red);
